Adds ?oType and related checking
Turns out the Terms as Links problem was real, and we weren't every detecting Terms as Terms in blobject-oriented queries. This adds an "oType" SPARQL variable to the "everything" query that can reliably be used, in combination with existing blank node values, to set a meaningful type on the octothorpe record. Closes #107